Changed the ncbi to hugo map (more coverage)

+++ data/gene_graphs.py
@@ -155,7 +155,5 @@ class HumanNetV2Graph(GeneInteractionGraph):
         self.nx_graph = nx.relabel.relabel_nodes(self.nx_graph, ncbi_to_hugo_map(self.nx_graph.nodes))
         # Remove nodes which are not covered by the map
         for node in list(self.nx_graph.nodes):
-            if isinstance(node, int):
+            if isinstance(node, float):
                 self.nx_graph.remove_node(node)
-
-
